fix: mypage client button
